# Avoid clearing Prettier cache when not using prettierrc