@BPScott 2020-06-12
Ooof that's a chunky speedup. Great investigating!

I've just tried this out with my [go-to codebase](https://github.com/shopify/polaris-react) that's a bit bigger (~700 files for eslint) and am seeing an 8 second speedup too - time to run just the prettier rule goes from ~16s to ~8s. Uncached total runtime goes from ~72s to ~64s. Our of interest what's the speedup for just the prettier rule on your codebase?

This is a big speedup so given most people are going to be spending time running lints rather than editing configs I'm rather tempted to say we straight up remove the clearCacheConfig call entirely and document that people are gonna need to reload their vscode windows if they edit their config.

VSCode eslint already requires you to reload the window to cachebust other plugin configs such as 
 changes to graphql schemas when using [eslint-plugin-graphql](https://github.com/microsoft/vscode-eslint/issues/477) so I don't think we should be a special case given how much we lose by trying to handle that.


Yo @dbaeumer, sorry for the wild ping, but it looks like vscode-eslint handles "if the .eslintrc changes then bust caches and reload eslint" [here](https://github.com/microsoft/vscode-eslint/blob/3e56a9eb7ecdc9d31d6aeb6a0d1ad799a9ef26ae/client/src/extension.ts#L894). It's not really vscode-eslint's problem but would you consider accepting adding `..prettierr{c.js,c.yaml,c.yml,c,c.json}` ([etc](https://prettier.io/docs/en/configuration.html)) to that dirty check for the sake of trying to do the right thing automagically in our case?