@dbaeumer 2020-06-15
@BPScott yes, I would accept suc a PR since there is a huge likelihood that if a prettier config file is present and eslint is running that the corresponding plugin is used.