@BPScott 2020-06-16
> Even if the vscode plugin does this, there area bunch of others that would also need to have this special case, like sublime, atom, and eslint_d.

Yepyep, and I'm trying to say that all of those also need some mechanism to solve e.g. cachebusting after graphql schema changes if consumers are using eslint-plugin-graphql. So I think the responsibility of specifying and handling that mechanism lies with the tools that are creating long-running eslint processes rather than the plugins themselves.